debug: introduce debug.{c,h}
Move debug-related functions here and introduce debug_dbg() which can be
used to replace if-debug-debug constructs.
